requirements.txt
Flask
python-dotenv
requests
supabase
google-generativeai